Handle Swarm-style prefixed names in Container.from_ps()
